Fixed ViewModel creation
